+++ apps/server/src/utils/validators.ts
@@ -29,3 +29,23 @@ export const withValidateAggregatorInPath =
 
     await handler(req, res);
   };
+
+export const withValidateAggregatorInQueryParams =
+  // eslint-disable-next-line @typescript-eslint/ban-types, @typescript-eslint/no-explicit-any
+  (handler: Function) => async (req: any, res: any) => {
+    const schema = Joi.object({
+      aggregator: createAggregatorValidator(),
+    });
+
+    const { error } = schema.validate({
+      aggregator: req.query.aggregator,
+    });
+
+    if (error) {
+      res.status(400);
+      res.send(he.encode(error.details[0].message));
+      return;
+    }
+
+    await handler(req, res);
+  };